backend: Move more API specific stuff into backend-api
