+++ packages/backend-api/package.json
@@ -15,13 +15,18 @@
     "@conversationai/moderator-backend-core": "1.1.0",
     "@conversationai/moderator-config": "1.1.0",
     "@conversationai/moderator-frontend-web": "1.1.0",
+    "@types/body-parser": "1.17.0",
+    "@types/compression": "0.0.36",
+    "@types/cors": "2.8.5",
     "@types/express": "4.16.0",
     "@types/express-ws": "3.0.0",
+    "@types/helmet": "0.0.43",
     "@types/joi": "10.4.0",
+    "@types/jsonwebtoken": "7.2.0",
     "@types/kue": "0.11.1",
     "@types/lodash": "4.14.132",
     "@types/mysql": "2.15.5",
-    "@types/passport": "0.4.5",
+    "@types/passport": "1.0.0",
     "@types/passport-google-oauth2": "0.1.2",
     "@types/passport-jwt": "3.0.1",
     "@types/qs": "6.5.1",
@@ -30,18 +35,23 @@
     "@types/request": "2.47.0",
     "@types/sequelize": "3.4.43",
     "@types/underscore.string": "0.0.30",
+    "body-parser": "1.19.0",
     "canvas": "2.2.0",
+    "compression": "1.7.4",
+    "cors": "2.8.5",
     "express": "4.16.4",
     "express-ws": "4.0.0",
     "googleapis": "32.0.0",
     "he": "1.2.0",
+    "helmet": "3.18.0",
     "joi": "10.5.2",
+    "jsonwebtoken": "8.5.1",
     "kue": "0.11.6",
     "lodash": "4.17.11",
     "moment": "2.21.0",
     "mysql": "2.15.0",
     "passport": "0.4.0",
-    "passport-google-oauth20": "1.0.0",
+    "passport-google-oauth20": "2.0.0",
     "passport-jwt": "4.0.0",
     "qs": "6.6.0",
     "randomstring": "1.1.5",
