+++ packages/backend-core/package.json
@@ -20,33 +20,21 @@
     "@conversationai/moderator-config": "1.1.0",
     "@types/joi": "10.4.0",
     "@types/js-yaml": "3.12.1",
-    "@types/jsonwebtoken": "7.2.0",
     "@types/lodash": "4.14.132",
     "@types/mysql": "2.15.5",
-    "@types/passport": "0.4.5",
-    "@types/passport-google-oauth2": "0.1.2",
-    "@types/passport-jwt": "3.0.1",
     "@types/request": "2.47.0",
     "@types/sequelize": "3.4.43",
     "@types/underscore.string": "0.0.30",
     "@types/yargs": "13.0.0",
-    "body-parser": "1.17.1",
-    "compression": "1.6.2",
-    "cors": "2.8.1",
     "express-winston": "2.5.1",
     "find-up": "3.0.0",
     "googleapis": "32.0.0",
-    "helmet": "3.5.0",
     "joi": "10.5.2",
-    "jsonwebtoken": "7.3.0",
     "js-yaml": "3.13.1",
     "lodash": "4.17.11",
     "moment": "2.21.0",
     "mysql": "2.15.0",
     "opentype.js": "0.6.9",
-    "passport": "0.4.0",
-    "passport-google-oauth20": "1.0.0",
-    "passport-jwt": "4.0.0",
     "request": "2.87.0",
     "sequelize": "3.30.4",
     "sequelize-cli": "2.7.0",
